moved openetlemetry service context into utilities
